# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ReWear API Documentation",
        "description": "ReWear - Sustainable Fashion Marketplace API\n *\n * **Mission:** Making fashion sustainable and accessible\n *\n * **Tagline:** Buy, Sell, Donate\n *\n * ## Features\n * - 🛍️ **Marketplace:** Buy and sell pre-loved clothing\n * - 💚 **Donations:** Donate items to verified charities\n * - 🚚 **Delivery:** Local driver network\n * - 🌍 **Impact:** Track environmental and social impact\n *\n * ## Authentication\n * All protected endpoints require a JWT Bearer token in the Authorization header:\n * ```\n * Authorization: Bearer YOUR_JWT_TOKEN\n * ```\n *\n * ## Rate Limiting\n * - **Registration OTP:** 5 requests per email per hour, 5 per IP per hour\n * - **Login:** Standard Laravel rate limiting\n * - **API Calls:** 60 requests per minute\n *\n * ## User Roles\n * - **user:** Regular buyer/seller (default)\n * - **charity:** Organization receiving donations (admin-created)\n * - **driver:** Verified delivery personnel (admin-approved)\n * - **admin:** Platform administrator\n *\n * ## Support\n * - **Email:** support@rewear.com\n * - **Phone:** +961 70 123456\n * ",
        "contact": {
            "name": "ReWear Support Team",
            "email": "support@rewear.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
...
        },
...
    ],
    "paths": {
...
        "/api/user/delete-account": {
            "delete": {
                "tags": [
                    "User Management"
                ],
                "summary": "Delete own account",
                "description": "Delete the authenticated user's own account",
                "operationId": "3da81b2bccb0536590d1fe0354a9dd2c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "MyPassword123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your account has been permanently deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid password"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/users/{userId}": {
            "delete": {
                "tags": [
                    "Admin - User Management"
                ],
                "summary": "Delete any user (Admin only)",
                "description": "Admin delete any user account",
                "operationId": "13884ba99396773eca983d62d85e5186",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Violated terms of service"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User account deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not an admin"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin - User Management"
                ],
                "summary": "Get all users (Admin only)",
                "description": "Get list of all users (Admin only)",
                "operationId": "b6e3df43263097b3605a503254c40efc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "user_type",
                        "in": "query",
                        "description": "Filter by user type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "user",
                                "charity"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users list retrieved successfully"
                    },
                    "403": {
                        "description": "Forbidden - Not an admin"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearer": []
        }
    ]
}